Dodana provjera je li user admin

+++ app/stories/routes.py
@@ -85,14 +85,18 @@ def validation(story_id):
     story_title = validationstory.title
     story_elements = sorted(StoryContent.query.filter_by(story_id=story_id), key=lambda x: x.ordinal_number)
     if request.method == 'POST':
-        if request.form['prihvPonudaPric'] == 'Prihvati':
-            validationstory.validated = True
-            db.session.add(validationstory)
-            db.session.commit()
-        elif request.form.get['odbPonudaPric'] == 'Odbaci':
-            db.session.delete(validationstory)
-            db.session.commit()
+        if current_user.id == 1:
+            if request.form['prihvPonudaPric'] == 'Prihvati':
+                validationstory.validated = True
+                db.session.add(validationstory)
+                db.session.commit()
+            elif request.form.get['odbPonudaPric'] == 'Odbaci':
+                db.session.delete(validationstory)
+                db.session.commit()
+        else:
+            abort(404)
     return render_template("prihvat_price.html", title="Prihvat price", story_title=story_title, story_elements=story_elements, elem_len=len(story_elements))
+
 
 @stories.route('/stories')
 def display_story_list():
